requirements.txt
Django==1.10.4
django-extensions==1.7.5
path.py==9.0
django-lazysignup==1.1.1
django-money==0.10.2
djangorestframework==3.6.2
